fixes cutout_to_black tests
