+++ django/inttest_to_black.cfg
@@ -1,3 +1,2 @@
 [unittest]
-test-file-pattern = int_test_to_black_view_uint8.py
-test-method-prefix = test_uint8_cutout_to_black
+test-file-pattern = int_test_to_black_view.py
